requirements.txt
numpy
scipy
matplotlib
torch
torchvision
skimage.metrics